refactor: output helpers
