+++ src/App.css
@@ -127,33 +127,9 @@ body{
   justify-content: space-around;
   border-radius: 5px;
   position: relative;
-  padding: 20px 0; /* rgba(233, 247, 239, 0.928); */
+  padding: 20px 0;
   background-color: rgba(148, 147, 144, 0.27);
 }
-
-.card::after{
-  content: '';
-  position: absolute;
-  width: 40px;
-  height: 80px;
-  top: 25%;
-  right: 50px;
-  background-color: rgba(255, 255, 255, 0.579);
-  border-radius: 50%;
-  transform: rotateZ(8deg);
-}
-
-/* .card::before{
-  content: '';
-  position: absolute;
-  width: 150px;
-  height: 150px;
-  top: 60%;
-  right: -90px;
-  background-color: white;
-  box-shadow: 1px 1px 10px white;
-  border-radius: 50%;
-} */
 
 .card:hover{
   cursor: pointer;
