+++ src/constants/colors.js
@@ -1,0 +1,8 @@
+export const colors = [
+    '#FC6B6E',
+    '#2196F3',
+    '#094BE8',
+    '#2196F3',
+    '#3ED1E0',
+    '#CF9B48',
+];
