+++ src/utils/searchIcon.js
@@ -1,0 +1,8 @@
+import { icons } from '../constants/icons';
+
+export const searchIcon = (search) => {
+    if (search) {
+        search = Object.entries(icons).filter((icon) => icon[0] === search);
+        return search[0][1];
+    }
+};
